takt: github-issue-413-refactor-acto
