+++ modules/actor/src/core/scheduler.rs
@@ -1,6 +1,5 @@
 //! Scheduler subsystem providing Pekko-compatible APIs.
 
-mod batch_mode;
 /// Cancellable entry, registry, and state types.
 pub mod cancellable;
 mod command;
@@ -21,7 +20,6 @@ mod mode;
 /// Scheduler policy types for periodic job execution.
 pub mod policy;
 mod runnable;
-mod runner_mode;
 mod scheduler_context;
 mod scheduler_core;
 mod scheduler_runner;
@@ -36,7 +34,6 @@ mod warning;
 #[cfg(test)]
 mod tests;
 
-pub use batch_mode::BatchMode;
 pub use command::SchedulerCommand;
 pub use config::SchedulerConfig;
 pub use delay_provider::SchedulerBackedDelayProvider;
@@ -44,15 +41,14 @@ pub use dispatcher_sender_shared::DispatcherSenderShared;
 pub use dump::SchedulerDump;
 pub use dump_job::SchedulerDumpJob;
 pub use error::SchedulerError;
-pub use execution_batch::ExecutionBatch;
+pub use execution_batch::{BatchMode, ExecutionBatch};
 pub use handle::SchedulerHandle;
 pub use metrics::SchedulerMetrics;
 pub use mode::SchedulerMode;
 pub use runnable::SchedulerRunnable;
-pub use runner_mode::RunnerMode;
 pub use scheduler_context::SchedulerContext;
 pub use scheduler_core::Scheduler;
-pub use scheduler_runner::SchedulerRunner;
+pub use scheduler_runner::{RunnerMode, SchedulerRunner};
 pub use scheduler_runner_owned::SchedulerRunnerOwned;
 pub use scheduler_shared::SchedulerShared;
 pub use warning::SchedulerWarning;
