+++ modules/actor/src/core/scheduler/task_run.rs
@@ -2,14 +2,11 @@
 
 mod task_run_entry;
 mod task_run_error;
-mod task_run_handle;
 mod task_run_on_close;
-mod task_run_priority;
 mod task_run_summary;
 
 pub(crate) use task_run_entry::{TaskRunEntry, TaskRunQueue};
+pub use task_run_entry::{TaskRunHandle, TaskRunPriority};
 pub use task_run_error::TaskRunError;
-pub use task_run_handle::TaskRunHandle;
 pub use task_run_on_close::TaskRunOnClose;
-pub use task_run_priority::TaskRunPriority;
 pub use task_run_summary::TaskRunSummary;
