+++ modules/actor/src/core/scheduler/tick_driver/auto_driver_metadata.rs
@@ -1,8 +1,22 @@
 //! Metadata describing auto-selected drivers.
 
+// Issue #413: AutoProfileKind は AutoDriverMetadata のフィールド型としてのみ使用されるため同居させる。
+#![allow(multiple_type_definitions)]
+
 use core::time::Duration;
 
-use super::{AutoProfileKind, TickDriverId};
+use super::TickDriverId;
+
+/// Classification of auto-detected driver profiles.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum AutoProfileKind {
+  /// Tokio runtime detected.
+  Tokio,
+  /// Embassy runtime detected.
+  Embassy,
+  /// Custom runtime.
+  Custom,
+}
 
 /// Metadata for automatic driver profile selection.
 #[derive(Debug, Clone)]
