--- modules/actor/src/core/scheduler/tick_driver/hardware_kind.rs
@@ -1,12 +1,0 @@
-//! Classification of hardware timer sources.
-
-/// Hardware timer source classification.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum HardwareKind {
-  /// Embassy-based timer.
-  Embassy,
-  /// SysTick-based timer.
-  SysTick,
-  /// Custom hardware timer.
-  Custom,
-}
